Allow a specific version or release of hab to be installed.
This change introduces a Dockerile build arg called `HAB_VERSION` which
can be used to set a specific version or version/release of hab from the
Depot.

The default build behavior is still in tack:

    docker build -t bscott/habitat .

But a specific version can be baked in with:

    docker build -t bscott/habitat --build-arg HAB_VERSION=0.8.0/20160729175817 .

This change should help to produce a tag-per-release model, tracking new
`core/hab` releases to the Depot.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>